requirements.txt
beautifulsoup4
certifi
chardet
click
colorama
Flask
idna
itsdangerous
Jinja2
lxml
MarkupSafe
pip
PyQt5
PyQt5-sip
PyQt5-stubs
PySocks
PyYAML
requests
pillow